requirements.txt
langchain_google_genai
streamlit
Pillow
ipykernel
python-dotenv
langchain_core